feat: scheduler refactor
