+++ crates/mate/Cargo.toml
@@ -23,7 +23,6 @@ anyhow = { workspace = true }
 async-trait = { workspace = true }
 bson = { workspace = true }
 clap = { workspace = true, features = ["derive", "env", "std"] }
-redis = { workspace = true, features = ["tokio-comp"] }
 tokio = { workspace = true, features = ["full"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
