+++ crates/scheduler/Cargo.toml
@@ -16,9 +16,11 @@ path = "src/bin/main.rs"
 anyhow = { workspace = true }
 async-trait = { workspace = true }
 clap = { workspace = true, features = ["derive", "env", "std"] }
+redis = { workspace = true, features = ["tokio-comp"] }
 tokio = { workspace = true, features = ["full"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 
 mate = { workspace = true }
 mate-fifo = { workspace = true }
+mate-proto = { workspace = true }
