+++ crates/scheduler/src/lib.rs
@@ -1,52 +1,55 @@
-use std::{sync::Arc, time::Duration};
+pub mod backend;
+pub mod task;
 
-use anyhow::Result;
-use tokio::{sync::mpsc::Sender, time::sleep};
+use std::time::{SystemTime, UNIX_EPOCH};
 
-use mate::scheduler::{backend::redis::RedisBackend, Scheduler};
+use anyhow::{Context, Result};
+use async_trait::async_trait;
 
-pub struct SchedulerTask {
-    scheduler: Arc<Scheduler<RedisBackend>>,
-    main_process_tx: Sender<String>,
+use mate_proto::{Job, JobId, PushJobDto};
+
+pub const SCHEDULER_JOB_PREFIX: &str = "mate:job";
+
+pub type Timestamp = u128;
+
+pub struct Scheduler<B: SchedulerBackend> {
+    backend: B,
 }
 
-impl SchedulerTask {
-    pub async fn new(
-        main_process_tx: Sender<String>,
-        scheduler: Arc<Scheduler<RedisBackend>>,
-    ) -> Result<Self> {
-        Ok(Self {
-            scheduler,
-            main_process_tx,
-        })
+impl<B: SchedulerBackend> Scheduler<B> {
+    pub fn new(backend: B) -> Self {
+        Self { backend }
     }
 
-    pub async fn run(&self) {
-        loop {
-            sleep(Duration::from_secs(1)).await;
-
-            // match scheduler.pop().await {
-            //     Ok(jobs) => {
-            //         if jobs.is_empty() {
-            //             continue;
-            //         }
-
-            //         if let Err(err) = main_pipe
-            //             .send(&Message::Text(format!("Jobs found: {}", jobs.len())))
-            //             .await
-            //         {
-            //             error!(%err, "Failed to send message to fifo");
-            //         }
-            //     }
-            //     Err(err) => {
-            //         if let Err(err) = main_pipe
-            //             .send(&Message::Text(format!("Failed to fetch jobs: {}", err)))
-            //             .await
-            //         {
-            //             error!(%err, "Failed to send message to fifo");
-            //         }
-            //     }
-            // }
-        }
+    #[inline]
+    pub async fn push(&self, job: PushJobDto) -> Result<JobId> {
+        self.backend.push(job).await
     }
+
+    #[inline]
+    pub async fn pop(&self) -> Result<Vec<Job>> {
+        self.backend.pop().await
+    }
+
+    #[inline]
+    pub async fn list(&self) -> Result<Vec<Job>> {
+        self.backend.list().await
+    }
+}
+
+#[async_trait]
+pub trait SchedulerBackend: Sized + Send + Sync + 'static {
+    type Config;
+
+    fn timestamp(&self) -> Result<Timestamp> {
+        SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .map(|dur| dur.as_millis())
+            .context("Failed to get current timestamp")
+    }
+
+    async fn new(config: Self::Config) -> Result<Self>;
+    async fn push(&self, job: PushJobDto) -> Result<JobId>;
+    async fn pop(&self) -> Result<Vec<Job>>;
+    async fn list(&self) -> Result<Vec<Job>>;
 }
